requirements/base.txt
dj-database-url==0.3.0
django-crispy-forms==1.4.0
django-kb>=0.2.0
django-secure==1.0
Markdown==2.4.1
Whoosh==2.6.0
